pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentspec"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "antlr4-python3-runtime==4.13",
    "jinja2>=3.1",
    "pydantic>=2.0",
    "langchain",
    "langchain-core",
    "langchain-text-splitters",
    "langchain_openai",
    "langchain_community",
    "langchain_experimental"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"
